Extract functions into handlers/utils
Refactored a little by extracting methods from the main package
into handlers, and some misc util definitions

Fixed in-cluster auth, which was omitted from the initial secret
creation PR due to a bad rebase
