+++ handlers/status.go
@@ -1,0 +1,7 @@
+package handlers
+
+import "net/http"
+
+func StatusHandler(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(http.StatusOK)
+}
